Adding the player class and the logic for updating the squares as we tap

+++ app/src/main/java/com/example/charlieyu/jump/Point.java
@@ -1,9 +1,9 @@
 package com.example.charlieyu.jump;
 
 public class Point {
-    double x,y;
+    float x,y;
 
-    public Point(double x, double y)
+    public Point(float x, float y)
     {
         this.x = x;
         this.y = y;
@@ -19,27 +19,27 @@ public class Point {
     static Point lineLineIntersection(Point A, Point B, Point C, Point D)
     {
         // Line AB represented as a1x + b1y = c1
-        double a1 = B.y - A.y;
-        double b1 = A.x - B.x;
-        double c1 = a1*(A.x) + b1*(A.y);
+        float a1 = B.y - A.y;
+        float b1 = A.x - B.x;
+        float c1 = a1*(A.x) + b1*(A.y);
 
         // Line CD represented as a2x + b2y = c2
-        double a2 = D.y - C.y;
-        double b2 = C.x - D.x;
-        double c2 = a2*(C.x)+ b2*(C.y);
+        float a2 = D.y - C.y;
+        float b2 = C.x - D.x;
+        float c2 = a2*(C.x)+ b2*(C.y);
 
-        double determinant = a1*b2 - a2*b1;
+        float determinant = a1*b2 - a2*b1;
 
         if (determinant == 0)
         {
             // The lines are parallel. This is simplified
             // by returning a pair of FLT_MAX
-            return new Point(Double.MAX_VALUE, Double.MAX_VALUE);
+            return new Point(Float.MAX_VALUE, Float.MAX_VALUE);
         }
         else
         {
-            double x = (b2*c1 - b1*c2)/determinant;
-            double y = (a1*c2 - a2*c1)/determinant;
+            float x = (b2*c1 - b1*c2)/determinant;
+            float y = (a1*c2 - a2*c1)/determinant;
             return new Point(x, y);
         }
     }
